my-app/conexion/conexionBD.py


# Importando Libreria mysql.connector para conectar Python con MySQL
import mysql.connector
from decouple import config


def connectionBD():
    try:
        # connection = mysql.connector.connect(
        connection = mysql.connector.connect(
            host=config('MYSQL_HOST'),
            user=config('MYSQL_USER'),
            passwd=config('MYSQL_PASSWORD'),
            database=config('MYSQL_DATABASE'),
            charset='utf8mb4',
            collation='utf8mb4_unicode_ci',
            raise_on_warnings=True

        )
        if connection.is_connected():
            # print("Conexión exitosa a la BD")
            return connection

    except mysql.connector.Error as error:
        print(f"No se pudo conectar: {error}")
