requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
openpyxl==3.1.2
protobuf==4.21.12
python-decouple==3.8
python-dotenv==1.0.1
Werkzeug==2.3.6
